refactor(tests): update namespace imports in shared infra extensions
- Changed `EnumerableExtensions` namespace to align with integration tests structure.
- Updated `AmazonDynamoDbExtensions` to use the updated `EnumerableExtensions` namespace.
